// src/app/interfaces/schedule.ts
// Generated by https://quicktype.io

export interface Schedule {
  id:         number;
  status:     string;
  start_date: string;
  end_date:   string;
  data:       Datum[];
  type:       string;
  image_url: string;
}

export interface Datum {
  id:           number;
  created_at:   Date;
  status?:       string;
  now?:          boolean;
  date?:         string;
  caption?:      string;
  ig_code?:      null;
  is_history?:   boolean;
  is_album?:     boolean;
  is_igtv?:      boolean;
  is_reels?:     boolean;
  ig_image_url?: null;
  type:         string;
  media_type?:   string;
  image?:        Image;
  channel:      Channel;
  socials?:      any[];
}

export interface Channel {
  id:          number;
  username:    string;
  profile_pic: string;
}

export interface Image {
  id:       number;
  filename: File;
  is_album: boolean;
  url:      string;
  type:     null;
}